backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.36
pydantic>=2.10.0
pydantic-settings>=2.7.0
httpx>=0.28.0
openai>=1.59.0
anthropic>=0.42.0
google-generativeai>=0.8.0
pyyaml>=6.0.2
python-dotenv>=1.0.1
jinja2>=3.1.4
pytest>=8.3.0
pytest-asyncio>=0.24.0
